requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
sqlalchemy==2.0.23
pyodbc==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.2
jupyter==1.0.0
ipykernel==6.27.1
